# src/gen/java/io/swagger/model/Car.java
/*
 * Taxi
 * Taxi management system API
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package io.swagger.model;

import com.fasterxml.jackson.annotation.JsonProperty;
import io.swagger.annotations.ApiModelProperty;

import javax.validation.constraints.NotNull;
import java.util.Objects;

/**
 * Car
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2018-01-17T10:00:02.220+03:00")
public class Car {
  @JsonProperty("id")
  private Integer id = null;
  
  @JsonProperty("color")
  private String color = null;
  
  @JsonProperty("plate")
  private String plate = null;
  
  public Car id(Integer id) {
    this.id = id;
    return this;
  }
  
  /**
   * Get id
   *
   * @return id
   **/
  @JsonProperty("id")
  @ApiModelProperty(value = "")
  public Integer getId() {
    return id;
  }
  
  public void setId(Integer id) {
    this.id = id;
  }
  
  public Car color(String color) {
    this.color = color;
    return this;
  }
  
  /**
   * Get color
   *
   * @return color
   **/
  @JsonProperty("color")
  @ApiModelProperty(required = true, value = "")
  @NotNull
  public String getColor() {
    return color;
  }
  
  public void setColor(String color) {
    this.color = color;
  }
  
  public Car plate(String plate) {
    this.plate = plate;
    return this;
  }
  
  /**
   * Get plate
   *
   * @return plate
   **/
  @JsonProperty("plate")
  @ApiModelProperty(required = true, value = "")
  @NotNull
  public String getPlate() {
    return plate;
  }
  
  public void setPlate(String plate) {
    this.plate = plate;
  }
  
  
  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Car car = (Car) o;
    return Objects.equals(this.id, car.id) &&
            Objects.equals(this.color, car.color) &&
            Objects.equals(this.plate, car.plate);
  }
  
  @Override
  public int hashCode() {
    return Objects.hash(id, color, plate);
  }
  
  
  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Car {\n");
    
    sb.append("    id: ").append(toIndentedString(id)).append("\n");
    sb.append("    color: ").append(toIndentedString(color)).append("\n");
    sb.append("    plate: ").append(toIndentedString(plate)).append("\n");
    sb.append("}");
    return sb.toString();
  }
  
  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}