# src/main/resources/open_api_spec.yml
swagger: '2.0'

info:
  version: "1.0.0"
  title: Taxi
  description: Taxi management system API
schemes:
  - http
  - https
host: localhost:8080
basePath: /taxi

paths:
  /cars:
    get:
      summary: Get all cars
      description: Get all cars
      parameters:
        - name: pageSize
          description: Number of cars per page
          in: path
          type: integer
          required: true
        - name: pageNumber
          description: Page number
          in: path
          type: integer
          required: true
      responses:
        200:
          description: All available cars
          schema:
            $ref: '#/definitions/Cars'
    post:
      summary: Create a car
      description: Create a car
      parameters:
        - name: car
          in: body
          description: A car
          schema:
            required:
              - color
              - plate
            properties:
              id:
                type: integer
              color:
                type: string
              plate:
                type: string
      responses:
        201:
          description: Created
        400:
          description: Couldn't create car
  /cars/{id}:
    get:
      summary: Get a car
      description: Get car by id
      parameters:
        - name: id
          description: Car identificator
          in: path
          required: true
          type: integer
      responses:
        200:
          description: A car
          schema:
            $ref: '#/definitions/Car'
        404:
          description: Not found
    delete:
      summary: Deletes a car
      description: Deletes car with id
      parameters:
        - name: id
          description: Car identificator
          in: path
          required: true
          type: integer
      responses:
        200:
          description: OK
        404:
          description: Not found



definitions:
  Car:
    required:
      - id
      - color
      - plate
    properties:
      id:
        type: integer
      color:
        type: string
      plate:
        type: string
  Cars:
    type: array
    items:
      $ref: '#/definitions/Car'